:sparkles: game 전체 검색, 키워드 검색, 카테고리 검새, 상세 정보 확인

+++ src/main/java/com/a404/boardgamers/Board/Service/BoardService.java
@@ -19,7 +19,6 @@ import org.springframework.stereotype.Service;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.transaction.Transactional;
-import java.awt.print.Pageable;
 import java.util.*;
 
 @RequiredArgsConstructor
@@ -31,43 +30,45 @@ public class BoardService {
 
     @Transactional
     public ResponseEntity uploadQuestion(BoardDTO.BoardUploadRequest uploadRequest, HttpServletRequest httpServletRequest) {
+//        System.out.println(uploadRequest.getWriterNickname() + " > " + uploadRequest.getTitle());
         String userId = TokenExtraction.check(httpServletRequest);
         Optional<User> user = userRepository.findUserById(userId);
+//        Optional<User> user = userRepository.findUserById(uploadRequest.getWriterNickname());
         if (!user.isPresent()) {
             return Response.newResult(HttpStatus.UNAUTHORIZED, "유효하지 않은 접근입니다.", null);
         }
+//        System.out.println(uploadRequest.getContent() + " >>> by " + uploadRequest.getWriterNickname());
         Board question = Board.builder()
                 .title(uploadRequest.getTitle())
                 .content(uploadRequest.getContent())
                 .writerId(user.get().getId())
                 .writerNickname(user.get().getNickname())
                 .build();
-        boardRepository.save(question);
-        return Response.newResult(HttpStatus.OK, "문의사항을 등록하였습니다.", question);
+        Board questionRequest = boardRepository.save(question);
+        return Response.newResult(HttpStatus.OK, "문의사항을 등록하였습니다.", questionRequest);
     }
 
     public ResponseEntity getQuestion(int id) {
-        Optional<Board> getQuestion = boardRepository.findBoardById(id);
-        String questionUploadedDate = TimestampToDateString.convertDate(getQuestion.get().getAddDate());
+        Optional<Board> question = boardRepository.findBoardById(id);
+        String questionUploadedDate = TimestampToDateString.convertDate(question.get().getAddDate());
         BoardDTO.BoardResponse questionResponse = BoardDTO.BoardResponse.builder()
-                .id(getQuestion.get().getId())
-                .title(getQuestion.get().getTitle())
-                .content(getQuestion.get().getContent())
+                .id(question.get().getId())
+                .title(question.get().getTitle())
+                .content(question.get().getContent())
                 .addDate(questionUploadedDate)
+                .viewCnt(question.get().getViewCnt())
                 .build();
+        // 조회수 반영 로직 고민해보기.
 
-        // 답변도 같이 줄 것
-//        long totalPageItemCnt = boardReplyRepository.countBoardRepliesByQnaId(id);
+        // 답변 추가하기.
         List<BoardReply> answers = boardReplyRepository.findBoardRepliesByQnaId(id);
         ArrayList<BoardReplyDTO.BoardReplyResponse> arr = new ArrayList<>();
 
-        // 관리자 답변 밖에 없으니 writer 별도 표기 하지 말까? 아니면 관리자라고 표기할까.
         for (BoardReply item : answers) {
             String itemDate = TimestampToDateString.convertDate(item.getAddDate());
             arr.add(BoardReplyDTO.BoardReplyResponse.builder()
                     .title(item.getTitle())
                     .content(item.getContent())
-                    .writerNickname(item.getWriterNickname())
                     .addDate(itemDate)
                     .build());
         }
@@ -89,7 +90,7 @@ public class BoardService {
             return Response.newResult(HttpStatus.UNAUTHORIZED, "유효하지 않은 접근입니다.", null);
         }
         Optional<Board> question = boardRepository.findBoardById(boardUpdateRequest.getId());
-//        Board question = boardRepository.findBoardById(boardUpdateRequest.getId()).orElseGet(Board::new);
+
         if (!question.isPresent()) {
             return Response.newResult(HttpStatus.BAD_REQUEST, "존재하지 않는 게시글입니다.", null);
         }
@@ -102,11 +103,23 @@ public class BoardService {
     }
 
     public ResponseEntity deleteQuestion(int id, HttpServletRequest httpServletRequest) {
+        String userId = TokenExtraction.check(httpServletRequest);
+        if (userId == null) {
+            return Response.newResult(HttpStatus.UNAUTHORIZED, "로그인 후 이용해주세요", null);
+        }
         Optional<Board> question = boardRepository.findBoardById(id);
         if (!question.isPresent()) {
-
+            return Response.newResult(HttpStatus.BAD_REQUEST, "존재하지 않는 게시글이므로 해당 요청을 수행할 수 없습니다.", null);
+        }
+        if (!userId.equals(question.get().getWriterId())) {
+            return Response.newResult(HttpStatus.BAD_REQUEST, "게시글 삭제 권한이 없습니다.", null);
         }
         boardRepository.delete(question.get());
+        // 달린 답글도 삭제?
+        List<BoardReply> replies = boardReplyRepository.findBoardRepliesByQnaId(question.get().getId());
+        for (BoardReply item : replies) {
+            boardReplyRepository.delete(item);
+        }
         return Response.newResult(HttpStatus.OK, "게시글이 삭제되었습니다.", null);
     }
 
@@ -118,11 +131,11 @@ public class BoardService {
         linkedHashMap.put("nowPage", page);
         linkedHashMap.put("nowPageSize", pageSize);
         PageRequest pageRequest = PageRequest.of(page - 1, pageSize);
-        List<Board> questionList = boardRepository.findAllQuestions((Pageable) pageRequest);
-        List<BoardDTO.BoardResponse> arr = new ArrayList<>();
+        List<Board> questionList = boardRepository.findAllQuestions(pageRequest);
+        List<BoardDTO.BoardListResponse> arr = new ArrayList<>();
         for (Board item : questionList) {
             String itemDate = TimestampToDateString.convertDate(item.getAddDate());
-            arr.add(BoardDTO.BoardResponse.builder()
+            arr.add(BoardDTO.BoardListResponse.builder()
                     .id(item.getId())
                     .title(item.getTitle())
                     .writerNickname(item.getWriterNickname())
@@ -132,7 +145,31 @@ public class BoardService {
         }
 
         linkedHashMap.put("questions", arr);
-        return Response.newResult(HttpStatus.OK, "문의글을 불러왔습니다.", arr);
+        return Response.newResult(HttpStatus.OK, "문의글을 불러왔습니다.", linkedHashMap);
     }
 
+    public ResponseEntity findQuestionsByKeyword(String keyword, int page, int pageSize) {
+        long totalItemCount = boardRepository.countBoardsByTitleContains(keyword);
+        HashMap<String, Object> linkedHashMap = new LinkedHashMap<>();
+        linkedHashMap.put("totalPageItemCnt", totalItemCount);
+        linkedHashMap.put("totalPage", ((totalItemCount - 1) / pageSize) + 1);
+        linkedHashMap.put("nowPage", page);
+        linkedHashMap.put("nowPageSize", pageSize);
+        PageRequest pageRequest = PageRequest.of(page - 1, pageSize);
+        List<Board> questionList = boardRepository.findBoardsByTitleContainsOrderByAddDate(keyword, pageRequest);
+
+        List<BoardDTO.BoardListResponse> arr = new ArrayList<>();
+        for (Board item : questionList) {
+            String itemDate = TimestampToDateString.convertDate(item.getAddDate());
+            arr.add(BoardDTO.BoardListResponse.builder()
+                    .id(item.getId())
+                    .title(item.getTitle())
+                    .writerNickname(item.getWriterNickname())
+                    .addDate(itemDate)
+                    .viewCnt(item.getViewCnt())
+                    .build());
+        }
+        linkedHashMap.put("questions", arr);
+        return Response.newResult(HttpStatus.OK, keyword + "로 문의글을 검색하였습니다.", linkedHashMap);
+    }
 }
